src/streamlit/license_plate/requirements.txt
streamlit
opencv-python-headless
Pillow
numpy
ultralytics
pytesseract
